Load PyInfo from rules_python (Attempt 2)
This is to facilitate upgrading of rules_python and other parts of Bazel.

More specifically, this makes the code compatible with both PyInfo from
rules_python and the Bazel builtin PyInfo. It does this by making rules
accept and produce both providers. This allows the code to be compatible
with a mixture of Bazel builtin `py_*` rules and rules_python `py_*` rules,
which make upgrading rules_python and Bazel easier in the future.

While I'd like to avoid this logic, I saw some errors about the wrong
PyInfo being accepted/produced. I wasn't able to track down the particular
edge, nor figure out if it was one we could even fix if found. Making it
compatible for both cases avoids the problem entirely.

Additional logic is present to handle two cases that will occur as part
of upgrading to Bazel 8 and higher:

* The builtin PyInfo can be an alias to rules_python PyInfo. This occurs
  when Bazel's autoloading is enabled. Because accepting/producing the
  same provider is an error, additional logic is needed to only use one.
* The builtin PyInfo can be None. This occurs occurs
  when Bazel's autoloading is disabled and builtin providers are also
  disabled (but not removed).
* The builtin PyInfo symbol is removed. This occurs in Bazel 9 and higher.

Reverts 99d7a393281333064926422d9396bdb6b0595cca

PiperOrigin-RevId: 771196494